+++ common/src/types/mainskill/mainskills/ingredient-draw-s.ts
@@ -3,7 +3,7 @@ import { Mainskill } from '../mainskill';
 
 export const IngredientDrawS = new (class extends Mainskill {
   name = 'Ingredient Draw S';
-  RP = [880, 1251, 1726, 2383, 3290, 4846, 5843]; // unsettled
+  RP = [880, 1251, 1726, 2383, 3290, 4846, 5843];
   ingredientAmounts = [5, 6, 8, 11, 13, 16, 18];
   image = 'ingredient_draw';
   description = (params: AmountParams) =>