+++ common/src/types/pokemon/skill-pokemon.ts
@@ -64,17 +64,19 @@ import {
   DreamShardMagnetS,
   DreamShardMagnetSRange,
   EnergizingCheerS,
+  EnergizingCheerSNuzzle,
   EnergyForEveryone,
+  EnergyForEveryoneBerryJuice,
   EnergyForEveryoneLunarBlessing,
   ExtraHelpfulS,
   HelperBoost,
   IngredientDrawS,
+  IngredientDrawSSuperLuck,
   IngredientMagnetS,
   IngredientMagnetSPlus,
   Metronome,
   TastyChanceS
 } from '../mainskill/mainskills';
-import { IngredientDrawSSuperLuck } from '../mainskill/mainskills/ingredient-draw-s-super-luck';
 
 import type { Pokemon } from './pokemon';
 
@@ -495,6 +497,25 @@ export const WOBBUFFET: Pokemon = createSkillSpecialist({
   skill: EnergizingCheerS
 });
 
+export const SHUCKLE: Pokemon = createSkillSpecialist({
+  name: 'SHUCKLE',
+  pokedexNumber: 213,
+  frequency: toSeconds(1, 0, 0),
+  ingredientPercentage: 20.5,
+  skillPercentage: 5.9,
+  berry: LUM,
+  genders: BALANCED_GENDER,
+  carrySize: 16,
+  previousEvolutions: 0,
+  remainingEvolutions: 0,
+  ingredients: {
+    a: PURE_OIL,
+    b: ROUSING_COFFEE,
+    c: HONEY
+  },
+  skill: EnergyForEveryoneBerryJuice
+});
+
 export const HERACROSS: Pokemon = createSkillSpecialist({
   name: 'HERACROSS',
   pokedexNumber: 214,
@@ -739,6 +760,25 @@ export const WYNAUT: Pokemon = preEvolvedPokemon(WOBBUFFET, {
   carrySize: 7
 });
 
+export const SPHEAL_HOLIDAY: Pokemon = createSkillSpecialist({
+  name: 'SPHEAL_HOLIDAY',
+  pokedexNumber: 363,
+  frequency: toSeconds(0, 55, 0),
+  ingredientPercentage: 21.4,
+  skillPercentage: 5.0,
+  berry: RAWST,
+  genders: BALANCED_GENDER,
+  carrySize: 20,
+  previousEvolutions: 0,
+  remainingEvolutions: 0,
+  ingredients: {
+    a: PURE_OIL,
+    b: BEAN_SAUSAGE,
+    c: WARMING_GINGER
+  },
+  skill: TastyChanceS
+});
+
 export const BONSLY: Pokemon = preEvolvedPokemon(SUDOWOODO, {
   name: 'BONSLY',
   pokedexNumber: 438,
@@ -970,6 +1010,25 @@ export const DEDENNE: Pokemon = createSkillSpecialist({
   skill: TastyChanceS
 });
 
+export const TOGEDEMARU: Pokemon = createSkillSpecialist({
+  name: 'TOGEDEMARU',
+  pokedexNumber: 777,
+  frequency: toSeconds(0, 45, 0),
+  ingredientPercentage: 16.9,
+  skillPercentage: 5.4,
+  berry: BELUE,
+  genders: BALANCED_GENDER,
+  carrySize: 18,
+  previousEvolutions: 0,
+  remainingEvolutions: 0,
+  ingredients: {
+    a: MOOMOO_MILK,
+    b: GLOSSY_AVOCADO,
+    c: SOOTHING_CACAO
+  },
+  skill: EnergizingCheerSNuzzle
+});
+
 export const MIMIKYU: Pokemon = createSkillSpecialist({
   name: 'MIMIKYU',
   pokedexNumber: 778,
@@ -1081,6 +1140,7 @@ export const OPTIMAL_SKILL_SPECIALISTS: Pokemon[] = [
   UMBREON,
   SLOWKING,
   WOBBUFFET,
+  SHUCKLE,
   HERACROSS,
   RAIKOU,
   ENTEI,
@@ -1091,6 +1151,7 @@ export const OPTIMAL_SKILL_SPECIALISTS: Pokemon[] = [
   PLUSLE,
   MINUN,
   SWALOT,
+  SPHEAL_HOLIDAY,
   DRIFBLIM,
   HONCHKROW,
   LUCARIO,
@@ -1104,6 +1165,7 @@ export const OPTIMAL_SKILL_SPECIALISTS: Pokemon[] = [
   BRAVIARY,
   SYLVEON,
   DEDENNE,
+  TOGEDEMARU,
   MIMIKYU,
   TOXTRICITY_AMPED,
   TOXTRICITY_LOW_KEY,